PTS: r2
The compatibility is accommodated again, since for the current checkout a compatibility with the "Direct Post and Hosted Fields" operating mode is not possible, it only works with "Hosted page".

For the new checkout (v4.1.1) that only supports versions greater than or equal to PrestaShop 1.7.6.X, yes, it is possible to support all modes of operation.